Remove an error when Theano process exit.

+++ theano/gof/compilelock.py
@@ -369,6 +369,8 @@ class Unlocker(object):
         # the same try/except block. The reason is that while the attempt to
         # remove the file may fail (e.g. because for some reason this file does
         # not exist), we still want to try and remove the directory.
+        if os is None:
+            return
         try:
             os.remove(os.path.join(self.tmp_dir, 'lock'))
         except Exception:
